+++ Cask/remindian.rb
@@ -1,6 +1,6 @@
 cask "remindian" do
-  version "5.8.0"
-  sha256 "f9d9e380dee991d439d7ab0ceaedb13b3a256b6f45d9e925032944887f2b750f"
+  version "5.8.1"
+  sha256 "671036a7e7d49fb5614f4fcf38755c5bc3997b8304352684017eeeee2e5225d0"
 
   url "https://github.com/Santofer/Remindian/releases/download/v#{version}/Remindian-v#{version}.dmg"
   name "Remindian"